A lot of progress done to packet ack and ack return handeling

+++ SimpleNetLib/SimpleNetLib/Packet/CorePacketComponents/ReturnAckComponent.hpp
@@ -7,7 +7,7 @@ class NET_LIB_EXPORT ReturnAckComponent : public PacketComponent
 public:
     explicit ReturnAckComponent();
 
-    int16_t ackIdentifier;
+    int32_t ackIdentifier;
 };
 
 inline ReturnAckComponent::ReturnAckComponent()
